+++ Game/Source/Objects/Shape.cpp
@@ -1,5 +1,5 @@
-#include"Math/Vector.h"
 #include "GamePCH.h"
+#include"Math/Vector.h"
 #include<iostream>
 
 vec4 shape_heart[] =
@@ -14,5 +14,5 @@ vec4 shape_heart[] =
      vec4(0.05f,0.2f,0,0),
      vec4(0.0f,0.05f,0,0)
  };
- int totalVerts_heart= std::size(shape_heart);
+ int totalVerts_heart= (int)std::size(shape_heart);
